pyproject.toml
[tool.poetry]
name = "pyhms_evaluator"
version = "0.1.0"
description = ""
authors = ["WojtAcht <wachtelik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
scipy = "^1.12.0"
matplotlib = "^3.8.3"
pyhms = {path = "../pyhms", develop = true}
pyade-python = "^1.1"
multiprocess = "^0.70.16"
smac = "^2.0.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"
pre-commit = "^3.6.2"
nbformat = "^5.10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
